changed folder structure and added relative paths
