+++ package.json
@@ -61,6 +61,7 @@
     "prettier-plugin-tailwindcss": "^0.2.1",
     "tailwindcss": "^3.2.4",
     "typescript": "^4.9.4",
-    "vite": "^4.0.4"
+    "vite": "^4.0.4",
+    "vite-tsconfig-paths": "^4.0.3"
   }
 }
