+++ src/components/playComponent/quiz/quizComponent/options.ts
