+++ src/components/playComponent/quiz/quizComponent/timerComponent.tsx
@@ -1,4 +1,4 @@
-import React, { useEffect, useState } from "react";
+import { useEffect, useState } from "react";
 import { MdTimer } from "react-icons/md";
 
 type Time = {
