+++ vite.config.ts
@@ -1,4 +1,6 @@
 import { defineConfig } from "vite";
+import tsconfigPaths from "vite-tsconfig-paths";
+
 import react from "@vitejs/plugin-react";
 import dns from "dns";
 
@@ -23,7 +25,7 @@ export default defineConfig({
       },
     },
   },
-  plugins: [react()],
+  plugins: [react(), tsconfigPaths()],
 
   server: {
     port: 3000,
